refactor: move llm interaction to a different class
