+++ Makefile
@@ -14,8 +14,8 @@ test:
 	@echo "Running tests..."
 	.venv/bin/python -m pytest
 
-generate-integ-data:
-	@echo "Generating integration test data..."
+generate-mock-test-data:
+	@echo "Generating mock test data..."
 	.venv/bin/python tests/data_generator.py
 
 ci: lint test
